Implement enforce_license parameter for knife.run
As a user of a community Chef distribution, I would like to be able to use knife
bootstrap without requiring the license check [1]. It's expected that that user
would not be installing a licensed version of Chef Infra Client, but instead a
community distribution.

This provides a similar feature as used in other places which makes it much
easier to disable the license check.

[1] https://gitlab.com/cinc-project/distribution/workstation/-/issues/3

Signed-off-by: Lance Albertson <lance@osuosl.org>