use openssl 1.1.1i for darwin/os X
We do not currently need FIPS support on mac_os_x, which
is why we ship 1.0.2x.  1.0.2x does not build nicely under arm64/m1;
shipping 1.1.1i allows us to take advantage of darwin/arm64 support
in that version of openssl, and ship a consistent version on both darwin
platforms.

This has no effect on the version shipped for other platforms.

Signed-off-by: Marc A. Paradise <marc.paradise@gmail.com>