Move most compliance validation to pre-run
Because it is important that when possible, the compliance run
get associated with the converge (via run-id) this PR updates
compliance mode to pre-validate for most common issues before the
converge occurs.

This is a change of behavior, in that previously we would
wait until it was time to send the report after the converge before
validating, which would cause the report to not get captured for current
converge (run-id) if config errors were present.

This also adds error numbers to each of the failure conditions we
detect, in order to simplify providing the operator help with resolving
the errors.

Resolves #11106 and #11105

Signed-off-by: Marc A. Paradise <marc.paradise@gmail.com>