Allow az vault name to be included in secret name
This modifies the `:azure_key_vault` fetcher so that it's possible
to fetch a secret by embedding the vault name in the secret name
instead of providing it in configuration. This continues down the path
of making secrets accessible with less typing and sane default
expectations.

Example:

```
file "/home/ubuntu/test2" do
  content secret(name: "test-chef-infra-secrets/test-secret-1", service: :azure_key_vault)
end
```

Specifying vault name via configuration is still supported, but if it is
specified in the secret name as well that will take precedence.

Signed-off-by: Marc A. Paradise <marc.paradise@gmail.com>