Add support for provider action description
Often actions are defined on the provider and not on the resource.  This
change updates Provider to support a description opt when declaring the
action, which lets us improve our generated documentation.

This also updates ResourceInspector to look at the provider if the
resource itself does not offer an action description for a given action.

Example:

action :foo, description: "bar" do
...
end

Signed-off-by: Marc A. Paradise <marc.paradise@gmail.com>