Fix failures in mac_user
After merging https://github.com/chef/chef/pull/11731 our functional tests started failing. It turns out we're not being defensive about data missing when we parse the dscl -> plist -> current resource. On my mac after creating a single user 'foo' with all defaults the shell data is nil as is the is_hidden data. The way we try to get the first item in the array fails on a method missing error for nilclass.

Signed-off-by: Tim Smith <tsmith@chef.io>